+++ src/primdirslistwnd.py
@@ -101,7 +101,7 @@ class PrimDirsListWnd(wx.ScrolledWindow):
 
 		self.fntMorinus = ImageFont.truetype(common.common.symbols, self.FONT_SIZE)
 		self.fntSymbol = ImageFont.truetype(common.common.symbols, int(3*self.FONT_SIZE/2))
-		self.fntAspects = ImageFont.truetype(common.common.symbols, 3*self.FONT_SIZE/4)
+		self.fntAspects = ImageFont.truetype(common.common.symbols, int(3*self.FONT_SIZE/4))
 		self.fntText = ImageFont.truetype(common.common.abc, self.FONT_SIZE)
 		self.clrs = (self.options.clrdomicil, self.options.clrexal, self.options.clrperegrin, self.options.clrcasus, self.options.clrexil)
