+++ src/transitmwnd.py
@@ -55,8 +55,8 @@ class TransitMonthWnd(commonwnd.CommonWnd):
 		self.SetVirtualSize((self.WIDTH, self.HEIGHT))
 
 		self.fntMorinus = ImageFont.truetype(common.common.symbols, self.FONT_SIZE)
-		self.fntAspects = ImageFont.truetype(common.common.symbols, 3*self.FONT_SIZE/4)
-		self.fntSigns = ImageFont.truetype(common.common.symbols, 3*self.FONT_SIZE/4)
+		self.fntAspects = ImageFont.truetype(common.common.symbols, int(3*self.FONT_SIZE/4))
+		self.fntSigns = ImageFont.truetype(common.common.symbols, int(3*self.FONT_SIZE/4))
 		self.fntText = ImageFont.truetype(common.common.abc, self.FONT_SIZE)
 		self.fntRText = ImageFont.truetype(common.common.abc, int(self.FONT_SIZE*3/4))
 		self.clrs = (self.options.clrdomicil, self.options.clrexal, self.options.clrperegrin, self.options.clrcasus, self.options.clrexil)
